requirements.txt
streamlit==1.22.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==0.27.0